requirements.txt
pytesseract
streamlit
transformers
Pillow
tesseract